refactor: Client 통신용 DTO 분리
