--- src/main/java/com/integrated/techhub/auth/dto/response/AccessTokenResponseDto.java
@@ -1,7 +1,0 @@
-package com.integrated.techhub.auth.dto.response;
-
-public record AccessTokenResponseDto (
-        String accessToken
-){
-
-}
